Adjust some examples to avoid variations in tests
GAP recently changed `AsList` for permutation groups, making it faster,
but as a result it now returns elements in a different order than `AsSet`.

But some of the tests here relied on `AsList(G)` always returning the same
result, and thus they started to fail. To avoid this, we just use `AsSet(G)`
which is guaranteed to give the same result in all GAP versions.